1. refs #6394

+++ hoot-core-test/src/test/cpp/hoot/core/algorithms/RdpWayGeneralizerTest.cpp
@@ -291,7 +291,7 @@ public:
     RdpWayGeneralizer generalizer(map, 0.1);
     generalizer.generalize(way);
 
-    CPPUNIT_ASSERT_EQUAL(197, map->getNodeMap().size());
+    CPPUNIT_ASSERT_EQUAL((size_t)197, map->getNodeMap().size());
     CPPUNIT_ASSERT_EQUAL((size_t)148, way->getNodeIds().size());
     const QString outFile =
       writeMap(map, "test-output/algorithms/RdpWayGeneralizerTest",
@@ -337,7 +337,7 @@ public:
     RdpWayGeneralizer generalizer(map, 0.1);
     generalizer.generalize(way);
 
-    CPPUNIT_ASSERT_EQUAL(197, map->getNodeMap().size());
+    CPPUNIT_ASSERT_EQUAL((size_t)197, map->getNodeMap().size());
     CPPUNIT_ASSERT_EQUAL((size_t)137, way->getNodeIds().size());
     const QString outFile =
       writeMap(map, "test-output/algorithms/RdpWayGeneralizerTest",
